+++ lib/rblade/compiler/compiles_statements.rb
@@ -1,4 +1,5 @@
 require "rblade/compiler/statements/compiles_conditionals"
+require "rblade/compiler/statements/compiles_helpers"
 require "rblade/compiler/statements/compiles_html_attributes"
 require "rblade/compiler/statements/compiles_inline_ruby"
 require "rblade/compiler/statements/compiles_loops"
@@ -68,6 +69,7 @@ module RBlade
       "case" => [CompilesConditionals, :compileCase],
       "checked" => [CompilesConditionals, :compileChecked],
       "class" => [CompilesHtmlAttributes, :compileClass],
+      "delete" => [CompilesHelpers, :compileDelete],
       "disabled" => [CompilesConditionals, :compileDisabled],
       "else" => [CompilesConditionals, :compileElse],
       "elsif" => [CompilesConditionals, :compileElsif],
@@ -95,15 +97,18 @@ module RBlade
       "for" => [CompilesLoops, :compileFor],
       "forelse" => [CompilesLoops, :compileForElse],
       "if" => [CompilesConditionals, :compileIf],
+      "method" => [CompilesHelpers, :compileMethod],
       "next" => [CompilesLoops, :compileNext],
       "nextif" => [CompilesLoops, :compileNextIf],
       "once" => [CompilesOnce, :compileOnce],
+      "patch" => [CompilesHelpers, :compilePatch],
       "prepend" => [CompilesStacks, :compilePrepend],
       "prependonce" => [CompilesOnce, :compilePrependOnce],
       "production" => [CompilesConditionals, :compileProduction],
       "props" => [CompilesProps, :compileProps],
       "push" => [CompilesStacks, :compilePush],
       "pushonce" => [CompilesOnce, :compilePushOnce],
+      "put" => [CompilesHelpers, :compilePut],
       "readonly" => [CompilesConditionals, :compileReadonly],
       "required" => [CompilesConditionals, :compileRequired],
       "ruby" => [CompilesInlineRuby, :compile],